pyproject.toml
[project]
name = "pumpfun-sniper"
version = "0.1.0"
description = "Solana Pump.fun token sniper using Helius Geyser Enhanced WebSockets"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "picows>=1.7.0",
    "uvloop>=0.21.0",
    "aiohttp[speedups]>=3.11.0",
    "msgspec>=0.19.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "numba>=0.60.0",
    "numpy>=2.0.0",
    "typer>=0.15.0",
]

[project.scripts]
sniper = "pumpfun_sniper.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pumpfun_sniper"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]
